fix haiku error and separate sidebar
