+++ treefamily/Frontend/components/Sidebar/Profile.js
@@ -1,0 +1,25 @@
+import React from "react";
+
+export default function Profile() {
+  return (
+    <div className="flex-shrink-0 flex bg-gray-700 p-4">
+      <a href="#" className="flex-shrink-0 w-full group block">
+        <div className="flex items-center">
+          <div>
+            <img
+              className="inline-block h-9 w-9 rounded-full"
+              src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
+              alt=""
+            />
+          </div>
+          <div className="ml-3">
+            <p className="text-sm font-medium text-white">Tom Cook</p>
+            <p className="text-xs font-medium text-gray-300 group-hover:text-gray-200">
+              View profile
+            </p>
+          </div>
+        </div>
+      </a>
+    </div>
+  );
+}
