+++ treefamily/Frontend/components/Sidebar/Sidebar.js
@@ -1,12 +1,7 @@
 /* This example requires Tailwind CSS v2.0+ */
-import {
-  CalendarIcon,
-  ChartBarIcon,
-  FolderIcon,
-  HomeIcon,
-  InboxIcon,
-  UsersIcon,
-} from "@heroicons/react/outline";
+import { For } from "react-haiku";
+import SidebarItem from "./SidebarItem";
+import Profile from "./Profile";
 
 const navigation = [
   { name: "Dashboard", href: "#", current: true },
@@ -26,9 +21,7 @@ const adminNavigation = [
   { name: "Reports", href: "#", current: false },
 ];
 
-function classNames(...classes) {
-  return classes.filter(Boolean).join(" ");
-}
+
 
 export default function Sidebar() {
   return (
@@ -39,48 +32,22 @@ export default function Sidebar() {
             <img
               className="h-8 w-auto"
               src="https://tailwindui.com/img/logos/workflow-logo-indigo-500-mark-white-text.svg"
-              alt="Workflow"    
+              alt="Workflow"
             />
           </div>
           <nav
             className="mt-5 flex-1 px-2 bg-gray-800 space-y-1"
             aria-label="Sidebar"
           >
-            {navigation.map((item) => (
-              <a
-                key={item.name}
-                href={item.href}
-                className={classNames(
-                  item.current
-                    ? "bg-gray-900 text-white"
-                    : "text-gray-300 hover:bg-gray-700 hover:text-white",
-                  "group flex items-center px-2 py-2 text-sm font-medium rounded-md"
-                )}
-              >
-                <span className="flex-1">{item.name}</span>
-              </a>
-            ))}
+            <For
+              each={navigation}
+              render={(item, index) => (
+                <SidebarItem item={item} />
+              )}
+            />
           </nav>
         </div>
-        <div className="flex-shrink-0 flex bg-gray-700 p-4">
-          <a href="#" className="flex-shrink-0 w-full group block">
-            <div className="flex items-center">
-              <div>
-                <img
-                  className="inline-block h-9 w-9 rounded-full"
-                  src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
-                  alt=""
-                />
-              </div>
-              <div className="ml-3">
-                <p className="text-sm font-medium text-white">Tom Cook</p>
-                <p className="text-xs font-medium text-gray-300 group-hover:text-gray-200">
-                  View profile
-                </p>
-              </div>
-            </div>
-          </a>
-        </div>
+        <Profile />        
       </div>
       <div className="bg-white">abcd</div>
     </div>
