+++ treefamily/Frontend/next.config.js
@@ -4,4 +4,7 @@ const nextConfig = {
   swcMinify: true,
 }
 
-module.exports = nextConfig
+module.exports = {
+  transpilePackages: ['react-haiku'],
+  nextConfig,
+}
